[select] chore: move src files into component folders
