+++ packages/select/src/common/_variables.scss
@@ -1,0 +1,7 @@
+// Copyright 2022 Palantir Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0.
+
+@import "~@blueprintjs/core/lib/scss/variables";
+
+$select-popover-max-height: $pt-grid-size * 30 !default;
+$select-popover-max-width: $pt-grid-size * 40 !default;
