+++ packages/select/src/components/_index.scss
@@ -1,6 +1,6 @@
 // Copyright 2017 Palantir Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
+@import "multi-select/multi-select";
 @import "omnibar/omnibar";
-@import "select/multi-select";
 @import "select/select";
