+++ packages/select/src/components/index.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
+export * from "./multi-select/multiSelect";
+export * from "./multi-select/multiSelect2";
 export * from "./omnibar/omnibar";
 export * from "./query-list/queryList";
-export * from "./select/multiSelect";
-export * from "./select/multiSelect2";
 export * from "./select/select";
 export * from "./select/select2";
-export * from "./select/suggest";
-export * from "./select/suggest2";
+export * from "./suggest/suggest";
+export * from "./suggest/suggest2";
