+++ packages/select/src/components/multi-select/_multi-select.scss
@@ -1,8 +1,8 @@
 // Copyright 2017 Palantir Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-@import "~@blueprintjs/core/src/common/variables";
-@import "../select/select";
+@import "~@blueprintjs/core/lib/scss/variables";
+@import "../../common/variables";
 
 .#{$ns}-multi-select {
   min-width: $pt-grid-size * 15;
