+++ packages/select/src/components/omnibar/_omnibar.scss
@@ -1,13 +1,12 @@
 // Copyright 2017 Palantir Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-@import "~@blueprintjs/colors/lib/scss/colors";
-@import "~@blueprintjs/core/src/common/variables";
+@import "~@blueprintjs/core/lib/scss/variables";
 @import "~@blueprintjs/core/src/common/react-transition";
 
 $omnibar-height: 60vh !default;
 $omnibar-width: $pt-grid-size * 50 !default;
-$omnibar-input-height: $pt-grid-size * 4;
+$omnibar-input-height: $pt-grid-size * 4 !default;
 
 .#{$ns}-omnibar {
   $omnibar-transition-props: (
