+++ packages/select/src/components/select/_select.scss
@@ -1,10 +1,8 @@
 // Copyright 2017 Palantir Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-@import "~@blueprintjs/core/src/common/variables";
-
-$select-popover-max-height: $pt-grid-size * 30 !default;
-$select-popover-max-width: $pt-grid-size * 40 !default;
+@import "~@blueprintjs/core/lib/scss/variables";
+@import "../../common/variables";
 
 .#{$ns}-select-popover {
   &.#{$ns}-select-match-target-width {
