+++ packages/select/test/suggestTests.tsx
@@ -23,7 +23,7 @@ import { InputGroup, IPopoverProps, Keys, MenuItem, Popover } from "@blueprintjs
 
 import { IFilm, renderFilm, TOP_100_FILMS } from "../../docs-app/src/common/films";
 import { IItemRendererProps, QueryList } from "../src";
-import { ISuggestProps, ISuggestState, Suggest } from "../src/components/select/suggest";
+import { ISuggestProps, ISuggestState, Suggest } from "../src/components/suggest/suggest";
 import { selectComponentSuite } from "./selectComponentSuite";
 
 describe("Suggest", () => {
